Finished DemoBoard.c
main demoboard is now complete. It has two main functions: readbuttons, that takes the inputs from the buttons every loop. And displayMatrix, which takes an array[9] of colors and displays them. 

There is an example function toggleButtons which takes a color, and assigns the colors to an array[9] based on previous inputs